Front/avatar e integração api

+++ backend/src/main/java/com/ft/trans/repository/LevelRepository.java
@@ -6,6 +6,6 @@ import com.ft.trans.entity.Level;
 public interface LevelRepository extends JpaRepository<Level, Long>
 {
     Level findByLevel(Integer level);
-    Level findTopByXpRequiredLessThanEqualOrderByXpRequiredDesc(Integer xp);
-    Level findTopByXpRequiredGreaterThanOrderByXpRequiredAsc(Integer xp);
+    Level findTopByXpRequiredLessThanEqualOrderByXpRequiredDesc(Long xp);
+    Level findTopByXpRequiredGreaterThanOrderByXpRequiredAsc(Long xp);
 }
